Authorization for deleting the post

+++ controller/api/v1/posts_api.js
@@ -19,12 +19,17 @@ module.exports.destroy = async function (req, res) {
     try {
         let post = await Post.findById(req.params.id);
 
-        post.remove();
-
-        await Comment.deleteMany({ post: req.params.id });
-        return res.json(200, {
-            message: "Post and associated comments deleted succesfully!"
-        });
+        if (post.user == req.user.id) {
+            post.remove();
+            await Comment.deleteMany({ post: req.params.id });
+            return res.json(200, {
+                message: "Post and associated comments deleted succesfully!"
+            });
+        } else {
+            return res.json(401,{
+                message: "You cannot delete this post!!"
+            });
+        }
     } catch (err) {
         if (err) {
             return res.json(500, {
